fix bug

+++ Driver.py
@@ -5,9 +5,10 @@ import PARAMETER
 import os
 
 if __name__ == '__main__':
-
-    os.remove(PARAMETER.MODEL)
-    os.remove(PARAMETER.RESULT)
+    if os.path.exists(PARAMETER.MODEL):
+        os.remove(PARAMETER.MODEL)
+    if os.path.exists(PARAMETER.RESULT):
+        os.remove(PARAMETER.RESULT)
 
     Read_file.allfiles(PARAMETER.PACKAGE_TRAIN)
